# MM-21368: Optimize MapStringsToQueryParams