--- a/store/sqlstore/utils.go
+++ b/store/sqlstore/utils.go
@@ -4,9 +4,7 @@
 package sqlstore
 
 import (
-	"bytes"
 	"database/sql"
-	"fmt"
 	"strconv"
 	"strings"
 
@@ -32,7 +30,7 @@ func sanitizeSearchTerm(term string, escapeChar string) string {
 // Converts a list of strings into a list of query parameters and a named parameter map that can
 // be used as part of a SQL query.
 func MapStringsToQueryParams(list []string, paramPrefix string) (string, map[string]interface{}) {
-	keys := bytes.Buffer{}
+	var keys strings.Builder
 	params := make(map[string]interface{}, len(list))
 	for i, entry := range list {
 		if keys.Len() > 0 {
@@ -44,7 +42,7 @@ func MapStringsToQueryParams(list []string, paramPrefix string) (string, map[str
 		params[key] = entry
 	}
 
-	return fmt.Sprintf("(%v)", keys.String()), params
+	return "(" + keys.String() + ")", params
 }
 
 // finalizeTransaction ensures a transaction is closed after use, rolling back if not already committed.